+++ packages/@expo/cli/src/export/embed/exportServer.ts
@@ -19,9 +19,10 @@ import { removeAsync } from '../../utils/dir';
 import { env } from '../../utils/env';
 import { CommandError } from '../../utils/errors';
 import { exportApiRoutesStandaloneAsync } from '../exportStaticAsync';
-import { copyPublicFolderAsync } from '../publicFolder';
-import { ExportAssetMap, persistMetroFilesAsync } from '../saveAssets';
-import { Options } from './resolveOptions';
+import { copyPublicFolderAsync, getPublicFolderPath } from '../publicFolder';
+import { persistMetroFilesAsync } from '../saveAssets';
+import type { ExportAssetMap } from '../saveAssets';
+import type { Options } from './resolveOptions';
 import {
   isExecutingFromXcodebuild,
   logInXcode,
@@ -65,7 +66,7 @@ export async function exportStandaloneServerAsync(
     apiRoutesOnly: true,
   });
 
-  const publicPath = path.resolve(projectRoot, env.EXPO_PUBLIC_FOLDER);
+  const publicPath = getPublicFolderPath(projectRoot);
 
   // Copy over public folder items
   await copyPublicFolderAsync(publicPath, serverOutput);